requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
streamlit
python-dotenv